refactor: fields param
