+++ src/teamster/code_locations/kipptaf/definitions.py
@@ -35,8 +35,11 @@ from teamster.code_locations.kipptaf import (
 )
 from teamster.core.resources import (
     BIGQUERY_RESOURCE,
+    DLT_RESOURCE,
     GCS_RESOURCE,
     GOOGLE_DRIVE_RESOURCE,
+    GOOGLE_FORMS_RESOURCE,
+    GOOGLE_SHEETS_RESOURCE,
     OVERGRAD_RESOURCE,
     SSH_COUCHDROP,
     get_dbt_cli_resource,
@@ -109,13 +112,13 @@ defs = Definitions(
         "db_bigquery": BIGQUERY_RESOURCE,
         "dbt_cli": get_dbt_cli_resource(DBT_PROJECT),
         "dds": resources.DIBELS_DATA_SYSTEM_RESOURCE,
-        "dlt": resources.DLT_RESOURCE,
+        "dlt": DLT_RESOURCE,
         "email": resources.OUTLOOK_RESOURCE,
         "gcs": GCS_RESOURCE,
         "google_directory": resources.GOOGLE_DIRECTORY_RESOURCE,
         "google_drive": GOOGLE_DRIVE_RESOURCE,
-        "google_forms": resources.GOOGLE_FORMS_RESOURCE,
-        "gsheets": resources.GOOGLE_SHEETS_RESOURCE,
+        "google_forms": GOOGLE_FORMS_RESOURCE,
+        "gsheets": GOOGLE_SHEETS_RESOURCE,
         "io_manager_gcs_avro": get_io_manager_gcs_avro(CODE_LOCATION),
         "io_manager_gcs_file": get_io_manager_gcs_file(CODE_LOCATION),
         "io_manager": get_io_manager_gcs_pickle(CODE_LOCATION),
