+++ src/teamster/code_locations/kipptaf/resources.py
@@ -1,6 +1,5 @@
 from dagster import EnvVar
 from dagster_airbyte import AirbyteCloudWorkspace
-from dagster_dlt import DagsterDltResource
 from dagster_shared import check
 
 from teamster.libraries.adp.workforce_manager.resources import (
@@ -12,8 +11,6 @@ from teamster.libraries.amplify.mclass.resources import MClassResource
 from teamster.libraries.coupa.resources import CoupaResource
 from teamster.libraries.email.resources import EmailResource
 from teamster.libraries.google.directory.resources import GoogleDirectoryResource
-from teamster.libraries.google.forms.resources import GoogleFormsResource
-from teamster.libraries.google.sheets.resources import GoogleSheetsResource
 from teamster.libraries.knowbe4.resources import KnowBe4Resource
 from teamster.libraries.ldap.resources import LdapResource
 from teamster.libraries.powerschool.enrollment.resources import (
@@ -58,8 +55,6 @@ DIBELS_DATA_SYSTEM_RESOURCE = DibelsDataSystemResource(
     username=EnvVar("AMPLIFY_DDS_USERNAME"), password=EnvVar("AMPLIFY_DDS_PASSWORD")
 )
 
-DLT_RESOURCE = DagsterDltResource()
-
 OUTLOOK_RESOURCE = EmailResource(
     host=EnvVar("OUTLOOK_HOST"),
     port=int(check.not_none(value=EnvVar("OUTLOOK_PORT").get_value())),
@@ -68,14 +63,10 @@ OUTLOOK_RESOURCE = EmailResource(
     chunk_size=450,
 )
 
-GOOGLE_FORMS_RESOURCE = GoogleFormsResource()
-
 GOOGLE_DIRECTORY_RESOURCE = GoogleDirectoryResource(
     customer_id=EnvVar("GOOGLE_WORKSPACE_CUSTOMER_ID"),
     delegated_account=EnvVar("GOOGLE_DIRECTORY_DELEGATED_ACCOUNT"),
 )
-
-GOOGLE_SHEETS_RESOURCE = GoogleSheetsResource()
 
 KNOWBE4_RESOURCE = KnowBe4Resource(
     api_key=EnvVar("KNOWBE4_API_KEY"), server="us", page_size=500
