+++ src/teamster/core/resources.py
@@ -2,6 +2,7 @@ import os
 
 from dagster import EnvVar
 from dagster_dbt import DbtCliResource
+from dagster_dlt import DagsterDltResource
 from dagster_gcp import BigQueryResource, GCSResource
 from dagster_shared import check
 
@@ -9,6 +10,8 @@ from teamster import GCS_PROJECT_NAME
 from teamster.core.io_managers.gcs import GCSIOManager
 from teamster.libraries.deanslist.resources import DeansListResource
 from teamster.libraries.google.drive.resources import GoogleDriveResource
+from teamster.libraries.google.forms.resources import GoogleFormsResource
+from teamster.libraries.google.sheets.resources import GoogleSheetsResource
 from teamster.libraries.overgrad.resources import OvergradResource
 from teamster.libraries.powerschool.sis.resources import PowerSchoolODBCResource
 from teamster.libraries.ssh.resources import SSHResource
@@ -84,7 +87,13 @@ DEANSLIST_RESOURCE = DeansListResource(
     api_key_map="/etc/secret-volume/deanslist_api_key_map_yaml",
 )
 
+DLT_RESOURCE = DagsterDltResource()
+
 GOOGLE_DRIVE_RESOURCE = GoogleDriveResource()
+
+GOOGLE_FORMS_RESOURCE = GoogleFormsResource()
+
+GOOGLE_SHEETS_RESOURCE = GoogleSheetsResource()
 
 OVERGRAD_RESOURCE = OvergradResource(api_key=EnvVar("OVERGRAD_API_KEY"), page_limit=100)
